removing WellStateFullyImplicitBlackoilDense as the master
it caused some problem in rebasing.
